docs: add MIT License to LICENSE.md
This commit adds the MIT License text to the LICENSE.md file in the root directory of the repository. The license grants permission to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the software, subject to certain conditions. This commit was made as part of the `docs` type of commit, indicating that it is a documentation update that does not affect the code itself.